# nexus/docs/api/openapi.json
{"openapi":"3.0.3","info":{"title":"Nexus Inventory API — Phase 1","version":"1.0.0","description":"Core REST endpoints for categories, products, and inventory. Auth via NextAuth session cookie."},"servers":[{"url":"/","description":"Same origin"}],"tags":[{"name":"Categories"},{"name":"Products"},{"name":"Inventory"},{"name":"Docs"}],"components":{"securitySchemes":{"sessionCookie":{"type":"apiKey","in":"cookie","name":"next-auth.session-token"}}},"security":[{"sessionCookie":[]}],"paths":{"/api/categories":{"get":{"tags":["Categories"],"summary":"List categories","parameters":[{"name":"organization","in":"query","schema":{"type":"string"}},{"name":"tree","in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"},"401":{"description":"Unauthorized"}}},"post":{"tags":["Categories"],"summary":"Create category","responses":{"201":{"description":"Created"},"409":{"description":"Duplicate"}}}},"/api/categories/{id}":{"get":{"tags":["Categories"],"summary":"Get category","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"put":{"tags":["Categories"],"summary":"Update category","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Categories"],"summary":"Delete category","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"}}}},"/api/products":{"get":{"tags":["Products"],"summary":"List/search products","parameters":[{"name":"page","in":"query","schema":{"type":"integer"}},{"name":"limit","in":"query","schema":{"type":"integer"}},{"name":"search","in":"query","schema":{"type":"string"}},{"name":"category","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"post":{"tags":["Products"],"summary":"Create product","responses":{"201":{"description":"Created"}}}},"/api/products/{id}":{"get":{"tags":["Products"],"summary":"Get product","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"put":{"tags":["Products"],"summary":"Update product","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"patch":{"tags":["Products"],"summary":"Partial update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Products"],"summary":"Delete product","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"}}}},"/api/inventory":{"get":{"tags":["Inventory"],"summary":"List stock levels","parameters":[{"name":"warehouse","in":"query","schema":{"type":"string"}},{"name":"product","in":"query","schema":{"type":"string"}},{"name":"lowStock","in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK"}}},"post":{"tags":["Inventory"],"summary":"Create inventory item","responses":{"201":{"description":"Created"}}}},"/api/inventory/{id}":{"get":{"tags":["Inventory"],"summary":"Get item","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"put":{"tags":["Inventory"],"summary":"Update item","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Inventory"],"summary":"Delete item","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"}}}},"/api/inventory/adjust":{"post":{"tags":["Inventory"],"summary":"Adjust stock quantity","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["id","change","reason"],"properties":{"id":{"type":"string"},"change":{"type":"integer"},"reason":{"type":"string"}}}}}},"responses":{"200":{"description":"Adjusted"}}}},"/api/inventory/stock-take":{"post":{"tags":["Inventory"],"summary":"Physical stock take","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["warehouseId","counts"],"properties":{"warehouseId":{"type":"string"},"counts":{"type":"array","items":{"type":"object"}}}}}}},"responses":{"200":{"description":"Results"}}}},"/api/docs":{"get":{"tags":["Docs"],"summary":"OpenAPI JSON or Swagger UI","parameters":[{"name":"format","in":"query","schema":{"type":"string","enum":["json","html"]}}],"security":[],"responses":{"200":{"description":"Docs"}}}}}}